Relocated base functionality to ClientBase. The Client extends the ClientBase where all base functionality and utilities live and are not directly accessible by the end user. The Client class is now more clean and represents the actual functionality used by the end user.
The response lists are now also filtered (again) since the element 
present (eg productList->product[]) is not usable for using the 
response as a new request. 

A new method is available prepareRequest for preparing a request for an
API call. This method will convert any object present in the request to
an array and if the second parameters is used, an element can be added
to the lists : e.g. : 
    $client->prepareRequest($request, [
        'productList' => 'product',
        'articleList' => 'article',
        'lineList' => 'line',
        'paymentList' => 'payment',
        'preparationList' => 'line',
        'preparationMethodList' => 'preparationMethod',
        'imageList' => 'image',
        'customFieldList' => 'customField',
        'sortOrderGroupList' => 'sortOrderGroup',
    ]);
Will add these required elements to the requested lists.

Also docblocks are added to each method.

Known bugs / TODO :
-dietRestrictionList->dietRestriction is not filtered from response
-allergenList->allergen is not filtered from response